+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Fixed
 - `rtcontrol`: Show a more useful error message when using `--from=<hash>`
   with a hash that doesn't exist
+- Send assigned SCGI headers over unix sockets
 
 ## [2.7.0] - 2023-02-11